separate out test function and refactor line sample
