+++ pdf2d/samples_test.go
@@ -1,34 +1,18 @@
+// See also test_test.go
+
 package pdf2d_test
 
 import (
 	"testing"
 
 	"github.com/stanim/draw2d"
-	"github.com/stanim/draw2d/pdf2d"
-
 	"github.com/stanim/draw2d.samples"
 	"github.com/stanim/draw2d.samples/android"
 	"github.com/stanim/draw2d.samples/frameimage"
 	"github.com/stanim/draw2d.samples/gopher"
 	"github.com/stanim/draw2d.samples/helloworld"
+	"github.com/stanim/draw2d.samples/line"
 )
-
-func test(t *testing.T, sample draw2d.Sample) {
-	// Initialize the graphic context on an RGBA image
-	dest := pdf2d.NewPdf("L", "mm", "A4")
-	gc := pdf2d.NewGraphicContext(dest)
-	// Draw Android logo
-	fn, err := sample(gc, "pdf")
-	if err != nil {
-		t.Errorf("Drawing %q failed: %v", fn, err)
-		return
-	}
-	// Save to png
-	err = pdf2d.SaveToPdfFile(fn, dest)
-	if err != nil {
-		t.Errorf("Saving %q failed: %v", fn, err)
-	}
-}
 
 func TestSampleAndroid(t *testing.T) {
 	test(t, android.Main)
@@ -48,4 +32,8 @@ func TestSampleHelloWorld(t *testing.T) {
 
 func TestSampleFrameImage(t *testing.T) {
 	test(t, frameimage.Main)
+}
+
+func TestSampleLine(t *testing.T) {
+	test(t, line.Main)
 }
