+++ samples_test.go
@@ -1,36 +1,18 @@
-// Package draw2d_test gives test coverage with the command:
-// go test -cover ./... | grep -v "no test"
+// See also test_test.go
+
 package draw2d_test
 
 import (
-	"image"
 	"testing"
 
 	"github.com/stanim/draw2d"
-
 	"github.com/stanim/draw2d.samples"
 	"github.com/stanim/draw2d.samples/android"
 	"github.com/stanim/draw2d.samples/frameimage"
 	"github.com/stanim/draw2d.samples/gopher"
 	"github.com/stanim/draw2d.samples/helloworld"
+	"github.com/stanim/draw2d.samples/line"
 )
-
-func test(t *testing.T, draw draw2d.Sample) {
-	// Initialize the graphic context on an RGBA image
-	dest := image.NewRGBA(image.Rect(0, 0, 297, 210.0))
-	gc := draw2d.NewGraphicContext(dest)
-	// Draw Android logo
-	fn, err := draw(gc, "png")
-	if err != nil {
-		t.Errorf("Drawing %q failed: %v", fn, err)
-		return
-	}
-	// Save to png
-	err = draw2d.SaveToPngFile(fn, dest)
-	if err != nil {
-		t.Errorf("Saving %q failed: %v", fn, err)
-	}
-}
 
 func TestSampleAndroid(t *testing.T) {
 	test(t, android.Main)
@@ -48,4 +30,8 @@ func TestSampleHelloWorld(t *testing.T) {
 
 func TestSampleFrameImage(t *testing.T) {
 	test(t, frameimage.Main)
+}
+
+func TestSampleLine(t *testing.T) {
+	test(t, line.Main)
 }
